+++ ui/utils/moveCursorPos.js
@@ -1,0 +1,3 @@
+export function moveCursorPos(x, y) {
+    process.stdout.cursorTo(x, y);
+}
